[refactor] AccountScreen: move 1st level screen to root folder
